requirements.txt
requests
streamlit
openai
python-dotenv
pymupdf
numpy
paddleocr
pandas